fix: add ENABLE_SCOPED_CUSTOM_ELEMENT_REGISTRY flag
