+++ packages/@lwc/features/src/types.ts
@@ -105,6 +105,14 @@ export interface FeatureFlagMap {
      *  - `Node.prototype.getRootNode`
      */
     ENABLE_LIGHT_GET_ROOT_NODE_PATCH: FeatureFlagValue;
+
+    /**
+     * Flag to enable scoped custom element registry (aka pivots). This patches the global custom elements registry
+     * to support having LWC components with the same tag name as third-party custom elements.
+     *
+     * If this flag is disabled, then LWC components with the same tag name as third-party custom elements may conflict.
+     */
+    ENABLE_SCOPED_CUSTOM_ELEMENT_REGISTRY: FeatureFlagValue;
 }
 
 export type FeatureFlagName = keyof FeatureFlagMap;
