+++ packages/@lwc/integration-karma/scripts/shared/options.js
@@ -18,11 +18,15 @@ const FORCE_NATIVE_SHADOW_MODE_FOR_TEST = Boolean(process.env.FORCE_NATIVE_SHADO
 const ENABLE_NATIVE_CUSTOM_ELEMENT_LIFECYCLE = Boolean(
     process.env.ENABLE_NATIVE_CUSTOM_ELEMENT_LIFECYCLE
 );
+const ENABLE_SCOPED_CUSTOM_ELEMENT_REGISTRY = Boolean(
+    process.env.ENABLE_SCOPED_CUSTOM_ELEMENT_REGISTRY
+);
 
 module.exports = {
     // Test configuration
     COMPAT,
     ENABLE_NATIVE_CUSTOM_ELEMENT_LIFECYCLE,
+    ENABLE_SCOPED_CUSTOM_ELEMENT_REGISTRY,
     FORCE_NATIVE_SHADOW_MODE_FOR_TEST,
     SYNTHETIC_SHADOW_ENABLED: !DISABLE_SYNTHETIC,
     GREP: process.env.GREP,
